requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
tensorflow==2.16.1
Pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4